refactor(app): split AppDelegate extensions into their own files
NSMenuItemValidation and NSWindowDelegate conformances move from inline
extensions in AppDelegate.swift to AppDelegate+MenuValidation.swift and
AppDelegate+WindowDelegate.swift respectively.

Three properties (dependencies, terminalApp, windows) and the WindowEntry
nested type are bumped from private to internal to support the cross-file
extensions. handleToggleStatusBar similarly bumps from private to internal
so the validator can reference its selector.
