+++ Bellith/App/AppDelegate+MenuValidation.swift
@@ -1,0 +1,10 @@
+import AppKit
+
+extension AppDelegate: NSMenuItemValidation {
+    func validateMenuItem(_ menuItem: NSMenuItem) -> Bool {
+        if menuItem.action == #selector(handleToggleStatusBar) {
+            menuItem.state = dependencies.settings.showStatusBar ? .on : .off
+        }
+        return true
+    }
+}
